refactor(estrutura): Reorganiza pacotes de Medico, Paciente e Endereco para o domain
