+++ src/main/java/com/vollmed/api/controllers/PacienteController.java
@@ -1,6 +1,6 @@
 package com.vollmed.api.controllers;
 
-import com.vollmed.api.paciente.*;
+import com.vollmed.api.domain.paciente.*;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
