+++ src/main/java/com/vollmed/api/domain/medico/DadosCadastroMedico.java
@@ -1,27 +1,30 @@
-package com.vollmed.api.paciente;
+package com.vollmed.api.domain.medico;
 
-import com.vollmed.api.endereco.DadosEndereco;
+import com.vollmed.api.domain.endereco.DadosEndereco;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
-import org.hibernate.validator.constraints.br.CPF;
 
-public record DadosCadastroPaciente(
+public record DadosCadastroMedico(
 
         @NotBlank
         String nome,
 
-        @NotBlank @Email
+        @NotBlank
+        @Email
         String email,
-
-        @NotBlank @CPF
-        @Pattern(regexp = "\\d{3}\\.\\d{3}\\.\\d{3}\\-\\d{2}")
-        String cpf,
 
         @NotBlank
         String telefone,
+
+        @NotBlank
+        @Pattern(regexp = "\\d{4,6}")
+        String crm,
+
+        @NotNull
+        Especialidade especialidade,
 
         @NotNull @Valid
         DadosEndereco endereco) {
