+++ src/main/java/com/vollmed/api/domain/medico/DadosListagemMedico.java
@@ -1,4 +1,4 @@
-package com.vollmed.api.medico;
+package com.vollmed.api.domain.medico;
 
 public record DadosListagemMedico (Long id,String nome, String email, String crm, Especialidade especialidade
 ){
