+++ src/main/java/com/vollmed/api/domain/medico/MedicoRepository.java
@@ -1,10 +1,8 @@
-package com.vollmed.api.medico;
+package com.vollmed.api.domain.medico;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
 
 public interface MedicoRepository extends JpaRepository <Medico, Long> {
