+++ src/main/java/com/vollmed/api/domain/paciente/DadosAtualizacaoPaciente.java
@@ -1,9 +1,9 @@
-package com.vollmed.api.medico;
+package com.vollmed.api.domain.paciente;
 
-import com.vollmed.api.endereco.DadosEndereco;
+import com.vollmed.api.domain.endereco.DadosEndereco;
 import jakarta.validation.constraints.NotNull;
 
-public record DadosAtualizacaoMedico(
+public record DadosAtualizacaoPaciente(
         @NotNull
         Long id,
         String nome,
