+++ src/main/java/com/vollmed/api/domain/paciente/DadosCadastroPaciente.java
@@ -1,30 +1,27 @@
-package com.vollmed.api.medico;
+package com.vollmed.api.domain.paciente;
 
-import com.vollmed.api.endereco.DadosEndereco;
+import com.vollmed.api.domain.endereco.DadosEndereco;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
+import org.hibernate.validator.constraints.br.CPF;
 
-public record DadosCadastroMedico(
+public record DadosCadastroPaciente(
 
         @NotBlank
         String nome,
 
-        @NotBlank
-        @Email
+        @NotBlank @Email
         String email,
+
+        @NotBlank @CPF
+        @Pattern(regexp = "\\d{3}\\.\\d{3}\\.\\d{3}\\-\\d{2}")
+        String cpf,
 
         @NotBlank
         String telefone,
-
-        @NotBlank
-        @Pattern(regexp = "\\d{4,6}")
-        String crm,
-
-        @NotNull
-        Especialidade especialidade,
 
         @NotNull @Valid
         DadosEndereco endereco) {
