+++ src/main/java/com/vollmed/api/domain/paciente/PacienteRepository.java
@@ -1,11 +1,9 @@
-package com.vollmed.api.paciente;
+package com.vollmed.api.domain.paciente;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 public interface PacienteRepository extends JpaRepository<Paciente, Long> {
